Move definition from <Kokkos_Core.hpp> to their own dedicated headers
